+++ CHANGELOG.md
@@ -8,8 +8,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - showError in Messages wrapper 
-- checkbox with basic functionality
-- mdTheme to checkbox
+- md-checkbox
+- mdTheme to checkbox, switch
+- md-switch
 
 ## 0.0.1 -